… eskip/parser.go